pyproject.toml
[project]
name = "bugsbugger"
version = "0.1.0"
description = "A Telegram bot that nags you about bills and deadlines until you pay attention"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot[job-queue]>=22.0",
    "aiosqlite>=0.20.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
